Revert "Configurable Prompts and new Models (#4)"
This reverts commit 5cb639b222848d0c4fcf9fb90b047fb3f2fff9ca.
